feat: 金额全链路改用 Decimal，修复浮点误差和 -0.00 显示
parse.py 直接从 CSV 字符串构造 Decimal（不经过 float，避免二进制精度
提前混入）；balance.py 用 Decimal 累加；report.py 格式化前先量化到
分并把量化后的负零归一化为正零；cli.py 的缺省金额也改用 Decimal。
补测试覆盖类型、0.1+0.2 精度、以及 -0.00/极小负数的显示归一化。

+++ ledger/parse.py
@@ -1,6 +1,7 @@
 """读交易记录 CSV。列：date,account,amount,memo（memo 可省略），第一行可以是表头。"""
 import csv
 from dataclasses import dataclass
+from decimal import Decimal
 
 HEADER = ["date", "account", "amount", "memo"]
 
@@ -9,7 +10,7 @@ HEADER = ["date", "account", "amount", "memo"]
 class Transaction:
     date: str
     account: str
-    amount: float
+    amount: Decimal
     memo: str = ""
 
 
@@ -18,7 +19,7 @@ def parse_row(row):
         raise ValueError(f"expected at least 3 columns, got {len(row)}: {row!r}")
     date, account, amount = (cell.strip() for cell in row[:3])
     memo = row[3].strip() if len(row) > 3 else ""
-    return Transaction(date, account, float(amount), memo)
+    return Transaction(date, account, Decimal(amount), memo)
 
 
 def parse_lines(lines):
